Moved HomePage navbar into a component to clean up Homepage file
